refactor: Legalページを再度公開 #21
