+++ app/src/app/(public)/legal/[slug]/page.tsx
